На UI добавлено поле с количеством желаний https://github.com/stas12312/wishlists/issues/31
