+++ src/frontend/src/models/IWish.ts
@@ -7,5 +7,6 @@ export interface IWish {
     created_at?: Dayjs;
     date: string;
     user_id: number;
+    wishes_count?: number;
     is_active?: boolean;
 }
