+++ src/frontend/src/pages/wishlist/Wishlists.tsx
@@ -16,6 +16,7 @@ import WishlistService from "../../services/WishlistService";
 import WishlistEditDialog from "./WishlistEditDialog";
 import {Dayjs} from "dayjs";
 import WishlistFooter from "./WishlistFooter";
+import {getNoun} from "../../helpers/Functions";
 
 interface IWishLists extends IWish {
     onItemsChange: Function
@@ -147,8 +148,10 @@ function Wishes(props: IWishLists) {
         uuid,
         user_id,
         date,
+        wishes_count,
         onItemsChange
     } = props;
+
     return (
         <Grid item
               justifyContent="center"
@@ -166,10 +169,18 @@ function Wishes(props: IWishLists) {
                         />
                     }
                     title={
+                    <div>
                         <Typography variant="subtitle1"
                                     sx={{fontWeight: 700}}>
                             {name}
                         </Typography>
+                        <Typography variant="subtitle2"
+                                    color="text.secondary">
+                            {wishes_count ?
+                                getNoun(wishes_count, 'подарок', 'подарка', 'подарков') :
+                                'Нет подарков'}
+                        </Typography>
+                    </div>
                     }
                 />
                 <CardContent>
@@ -208,6 +219,7 @@ function Wishlists(props: IWishList) {
                                 date={wishlist.date}
                                 created_at={wishlist.created_at}
                                 key={wishlist.uuid}
+                                wishes_count={wishlist.wishes_count}
                                 onItemsChange={onItemsChange}/>
 
                     )
